Arrange and connect each js files
