+++ .prettierrc
@@ -1,3 +1,3 @@
 # @format
 
-{ "singleQuote": true }
+{ 'singleQuote': true }
