+++ modules/local_storage.js
@@ -1,6 +1,5 @@
 export const setLocalStorage = (newBooks) =>
   localStorage.setItem('books', JSON.stringify(newBooks));
 
-export const getFromLocalStorage = () => {
-  this.books = JSON.parse(localStorage.getItem('books')) ?? [];
-};
+export const getFromLocalStorage = () =>
+  JSON.parse(localStorage.getItem('books')) ?? [];
