+++ modules/selector.js
@@ -9,19 +9,3 @@ export const list = document.querySelector('.list');
 export const addNew = document.querySelector('.add');
 export const contact = document.querySelector('.contact');
 export const dateTime = document.querySelector('.date');
-
-list.addEventListener('click', () => {
-  bookList.classList.remove('hidden');
-  addContainer.classList.add('hidden');
-  contactContainer.classList.add('hidden');
-});
-addNew.addEventListener('click', () => {
-  bookList.classList.add('hidden');
-  addContainer.classList.remove('hidden');
-  contactContainer.classList.add('hidden');
-});
-contact.addEventListener('click', () => {
-  bookList.classList.add('hidden');
-  addContainer.classList.add('hidden');
-  contactContainer.classList.remove('hidden');
-});
